requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Vector database
chromadb==0.4.24

# Embeddings / NLP
sentence-transformers==2.6.1
torch>=2.0.0
numpy>=1.23.0

# Data handling
pydantic==2.7.0
scikit-learn>=1.3.0

# Optional but useful
python-dotenv==1.0.1
langchain
langchain-community

#""" if you wanna use ollama then install 
#ollama """
langchain_anthropic
langchain_google_genai
langchain_openai
